requirements.txt
sentence-transformers==3.0.0
faiss-cpu==1.8.0
markdown==3.6
tqdm==4.66.5
numpy==1.26.4
openai==1.40.0
python-dotenv==1.0.1
PyYAML==6.0.2
markitdown==0.1.3
rich==13.7.1
